[FIX] Use the correct scores in the initial column/row of each block.
The first cell of the column vector in each block row is not computed but set directly from the last cell of the current block row vector.
However, the row vector was modified to only store the previous diagonal.
This fix temporarily reuses the first 'empty' cell of the row vector to get the score from last cell of the previous column vector.