[FEATURE] Enable larger block sizes in saturated simd mode.
The block size is now maximised to fully use the value range of 256 values on the saturated int8_t type.
The larger block size induces an offset of the zero position, which is added to the saturated dp vector and the saturated dp template.
Also the saturated dp tracker had to be modified slightly.