Adapt traversal policy to compute column slice with a profile.
The row sequence slice for the current column slice of the processed DP matrix tile is converted to a profile score, which is then returned instead of the original sequence slice to the core DP algorithm.
Ths substitution score then just queries the profile for the given sequence character read from the column sequence.