Fix saturated profile matrix alignment using SIMD instructions.
This fix uses the profile lane adaptor to convert for each column slice the respective row sequence slice to a profile score.
This makes the test work again. However, the current version needs refactoring by populating the profile transformation up such that it is only done for the current saturated column at the beginning.