Switches from static array to vector to store profiles in profile scoring scheme.
The SIMD profile for each character of the alphabet was stored inside of a std::array requirering that the width of the transformed slice over the row sequence must be known at compile time.
This commit changes the underlying storage model to dynamically allocated memory using std::vector instead. Thus, the size for each character profile can be determined from the passed row slice to be converted.